+++ src/include/AmazonAbstractWebServiceClient.h
@@ -95,7 +95,7 @@
 -(AmazonServiceResponse *)invoke:(AmazonServiceRequest *)generatedRequest rawRequest:(AmazonServiceRequestConfig *)originalRequest unmarshallerDelegate:(Class)unmarshallerDelegate andSign:(BOOL)signRequest;
 
 
--(void)pauseExponentially:(int)tryCount;
+-(void)pauseExponentially:(NSInteger)tryCount;
 -(BOOL)shouldRetry:(AmazonServiceResponse *)response;
 -(BOOL)shouldRetry:(AmazonServiceResponse *)response exception:(NSException *)theException;
